remove more dead code after StaticAnalyzer plugin refactor
Note that somehow this causes test regressions, but only when
the static analyzer libraries are linked statically with clang.
IE, it works when configured with `cmake -DUSE_SHARED_LIBS`.
I guess the linker is optimizing it away even though it contains
a static constructor with side-effects...

The error message (err_fe_invalid_plugin_name):
"(frontend): unable to find plugin 'analyzer'"